Modified round number to be Int instead of Long in BeaconBlockHeader (#69)

+++ maru/consensus/src/main/kotlin/maru/consensus/qbft/adapters/QbftBlockInterfaceAdapter.kt
@@ -30,7 +30,7 @@ class QbftBlockInterfaceAdapter : QbftBlockInterface {
     val beaconBlockHeader = proposalBlock.header.toBeaconBlockHeader()
     val replacedBeaconBlockHeader =
       beaconBlockHeader.copy(
-        round = roundNumber.toULong(),
+        round = roundNumber.toUInt(),
       )
     return QbftBlockAdapter(
       BeaconBlock(replacedBeaconBlockHeader, proposalBlock.toBeaconBlock().beaconBlockBody),
